+++ zopyx/existdb/tests/base.py
@@ -29,6 +29,7 @@ import plone.app.dexterity
 EXIST_DB_URL = os.environ.get('EXIST_DB_URL', 'http://localhost:6080')
 EXIST_DB_USERNAME = os.environ.get('EXIST_DB_USERNAME', 'admin')
 EXIST_DB_PASSWORD = os.environ.get('EXIST_DB_PASSWORD', 'admin')
+EXIST_DB_EMULATION = os.environ.get('EXIST_DB_EMULATION', 'existdb')
 
 
 class PolicyFixture(PloneSandboxLayer):
@@ -37,7 +38,7 @@ class PolicyFixture(PloneSandboxLayer):
 
     def setUpZope(self, app, configurationContext):
 
-#        xmlconfig.file('meta.zcml', z3c.jbot, context=configurationContext)
+        #        xmlconfig.file('meta.zcml', z3c.jbot, context=configurationContext)
         for mod in [plone.app.dexterity,
                     zopyx.existdb,
                     ]:
@@ -62,11 +63,12 @@ class PolicyFixture(PloneSandboxLayer):
         settings.existdb_username = unicode(EXIST_DB_USERNAME)
         settings.existdb_password = unicode(EXIST_DB_PASSWORD)
         settings.existdb_url = unicode(EXIST_DB_URL)
+        settings.existdb_emulation = unicode(EXIST_DB_EMULATION)
 
         self.connector = plone.api.content.create(
-                container=portal, 
-                type='zopyx.existdb.connector', 
-                id='connector')
+            container=portal,
+            type='zopyx.existdb.connector',
+            id='connector')
 
     def tearDownZope(self, app):
         # Uninstall product
@@ -90,4 +92,3 @@ class TestBase(unittest2.TestCase):
         """ Login as manager """
         user = self.portal.acl_users.getUser(uid)
         newSecurityManager(None, user.__of__(self.portal.acl_users))
-