+++ docs/source/HISTORY.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+0.2.9 (2014-11-01)
+------------------
+- support for overriding credentials locally 
+
+0.2.8 (2014-11-01)
+------------------
+- minor fix for mounting Plone sites over WebDAV into another Plone site
+
 0.2.7 (2014-11-01)
 ------------------
 - experimental support for BaseX XML database through the WebDAV API.